Fix start crash if file manager plugin is disabled
The file manager plugin already created widgets while it was temporarily
loaded for getting its attributes on application start. The new TreeView
widget then caused a crash later because its implementation was unloaded
with the deactivated plugin. This was fixed by now creating the widgets
in the plugin's init() method which is only called for enabled plugins.